move pyfiles into unit test fixture
This makes the test harness DRYer and allows any unit test
to access the list of python files.
