+++ test/unit/fixture.py
@@ -15,3 +15,12 @@ class CleanerBotTestFixture(unittest.TestCase):
         print
         print("Testing in: %s" % REPO_DIR)
         print
+        self.files = self.find_py_files(REPO_DIR)
+
+    def find_py_files(self, dirname):
+        pyfiles = []
+        for root, dirs, files in os.walk(dirname):
+            for file in files:
+                if file.endswith('.py'):
+                    pyfiles.append(os.path.join(root, file))
+        return pyfiles
